% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_nonlinearities}
\alias{apc_nonlinearities}
\title{Extract non-linear APC effects}
\usage{
apc_nonlinearities(apc, intercept = FALSE)
}
\arguments{
\item{apc}{APC model object from \code{apc()}}

\item{intercept}{Logical; include intercept in output}
}
\value{
List of non-linear effect estimates
}
\description{
Extract non-linear APC effects
}
