% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_plot_nonlinearities}
\alias{apc_plot_nonlinearities}
\title{Plot non-linear APC effects}
\usage{
apc_plot_nonlinearities(model)
}
\arguments{
\item{model}{APC model object from \code{apc()}}
}
\value{
ggplot2 object showing non-linear effects
}
\description{
Plot non-linear APC effects
}
