% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_plot_two2d}
\alias{apc_plot_two2d}
\title{Plot APC two-dimensional visualization}
\usage{
apc_plot_two2d(apc)
}
\arguments{
\item{apc}{APC model object from \code{apc()}}
}
\value{
ggplot2 object
}
\description{
Plot APC two-dimensional visualization
}
