% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gssrac}
\alias{gssrac}
\title{GSS}
\format{
A data frame with 5,234 rows and 5 variables:
\describe{
  \item{rac}{X}
  \item{year}{X}
  \item{age}{X}
  \item{cohort}{X}
  \item{region}{X}
  \item{sex}{X}
  \item{wtssall}{X}
}
}
\usage{
gssrac
}
\description{
GSS
}
\keyword{datasets}
