% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_events.R
\name{plot.decompose_events}
\alias{plot.decompose_events}
\title{Plot a decompose_events object}
\usage{
\method{plot}{decompose_events}(x, ...)
}
\arguments{
\item{x}{A `decompose_events` object returned by [decompose_events()].}

\item{...}{Not used.}
}
\value{
A `ggplot` object.
}
\description{
Plots observed pairwise outcome changes (as line segments by event type)
alongside implied counterfactual trajectories.
}
