% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{print.apc_model}
\alias{print.apc_model}
\title{Print an APC model}
\usage{
\method{print}{apc_model}(x, ...)
}
\arguments{
\item{x}{An `apc_model` object returned by [apc()].}

\item{...}{Not used.}
}
\value{
`x`, invisibly.
}
\description{
Print an APC model
}
