% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_events.R
\name{print.decompose_events}
\alias{print.decompose_events}
\title{Print a decompose_events object}
\usage{
\method{print}{decompose_events}(x, ...)
}
\arguments{
\item{x}{A `decompose_events` object returned by [decompose_events()].}

\item{...}{Not used.}
}
\value{
`x`, invisibly.
}
\description{
Print a decompose_events object
}
