% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_plot_total}
\alias{apc_plot_total}
\title{Plot total APC effects under assumption}
\usage{
apc_plot_total(model, assumption)
}
\arguments{
\item{model}{APC model object from \code{apc()}}

\item{assumption}{Character specifying linear trend assumption}
}
\value{
ggplot2 object showing total effects
}
\description{
Plot total APC effects under assumption
}
