% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_ic.R
\name{cr_ic}
\alias{cr_ic}
\title{Desc}
\usage{
cr_ic(data, formula, weight = NULL, model = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{X}

\item{weight}{X}

\item{model}{Can be a model object that supports `predict` or a model formula.}
}
\value{
Returns ...
}
\description{
Desc
}
\examples{
# todo
}
\references{
X
}
