% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_membership}
\alias{eu_membership}
\title{EU Membership data}
\format{
A data frame with 29 rows and 3 variables:
\describe{
  \item{country}{X}
  \item{date}{X}
  \item{event_type}{X}
}
}
\usage{
eu_membership
}
\description{
EU Membership data
}
\keyword{datasets}
