% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality_us}
\alias{mortality_us}
\title{US mortality rates, 1933--2019}
\format{
A data frame with 15,660 rows and 4 variables:
\describe{
  \item{year}{Year.}
  \item{age}{Age.}
  \item{sex}{Sex, either 'male' or 'female'.}
  \item{death_rate}{Central death rate m(x): deaths divided by exposure (HMD `Mx_1x1`).
    Not a probability -- it can exceed 1 at the oldest ages.}
}
}
\source{
HMD. Human Mortality Database.
 Max Planck Institute for Demographic Research (Germany),
 University of California, Berkeley (USA),
 and French Institute for Demographic Studies (France). Available at www.mortality.org.
}
\usage{
mortality_us
}
\description{
Mortality rates by sex and age. Used in examples for
\code{\link{decompose_aggregated}}.
}
\keyword{datasets}
