% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_ic.R
\name{plot.cr_ic_decomposition}
\alias{plot.cr_ic_decomposition}
\title{Desc}
\usage{
\method{plot}{cr_ic_decomposition}(x, total = TRUE, methods = NULL, ...)
}
\arguments{
\item{x}{X}

\item{total}{X}

\item{methods}{X}

\item{...}{Not used.}
}
\value{
Returns ...
}
\description{
Desc
}
\examples{
# todo
}
\references{
X
}
