% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpp_data}
\alias{wpp_data}
\title{WPP World Population data}
\format{
A data frame with 43,186 rows and 3 variables:
\describe{
  \item{Location}{X}
  \item{Time}{X}
  \item{PopTotal}{X}
}
}
\usage{
wpp_data
}
\description{
WPP
}
\keyword{datasets}
